Fixes #17 Warning on tagged enum variants
When an enum was tagged with #[attr_args], a warning is generated for
each variant of the tagged enum. This was due to inner functions of the
generated code having non snake case names. The issue was resolved by
adding a #[allow(non_snake_case)] to each of the generated functions.
Another possible solution would be to convert the camel case names to
snake case names, but disabling the lint on each of the functions was
faster.