Fixes #18 Adds support for array parameter values
Attribution now supports parameters whose value is a collection of
parameters. The type used in the struct to support this behavior is a
Vec whose element types are any other supported parameter value type
(e.g. Vec<i64>).

As part of this effort, the code used to parse the attribute input
went through a major overhaul. This included implementing the 'Parse'
trait for the 'ParamVal' enum and no longer using the 'MetaList' from
the syn crate. These changes to the parsing logic will grant more
flexibility in the future for the types of syntax that can be parsed as
attribute input.

As part of the implementation of this feature, a new variant to the
'ParamVal' enum was also added. The 'Array' variant stores the value as
a Vec<ParamVal>. Theoretically this means that heterogenous array values
could be used in the future.

The following is an example of attribute args and the corresponding
usage of the attribute. NOTE: Pound signs omitted due to commit message
constraints.

[attr_args]
struct AttrArgs {
    vals: Vec<i64>
}

// ...

[my_attr(vals = [1, 2, 3])]
fn foo() {
    //...
}

As part of this effort, a new example crate was added that demonstrates
the usage of multiple types of parameters.